improve topcat intro message generation
Generate the intro message at container build time and write it
into a file.  Then cause the contents of that file to be output
by /etc/bash.bashrc, which hopefully means the message will get
written to the container xterm on startup.
The previous behaviour was writing content to a stdout that was
not attached to the xterm.