replace URL for topcat container with versioned one
Previously the versioned URL for topcat release JARs was FTP-only,
which Docker didn't seem to like, so the ADD URL used the
location of the most recent release.  However historical versions
have now been made available over HTTP, so use the versioned
URL for topcat 4.8 instead.

Since the content of the URL is the same as when the container was
built, this change shouldn't make a material change to the
container content.

Note that v4.8 is no longer the most recent release - v4.8-1 is now
available and includes some bugfixes and enhancements so it could
be updated and rebuilt.